+++ lib/plug_rest/router.ex
@@ -62,29 +62,15 @@ defmodule PlugRest.Router do
   A request to "/hello/value/extra" will populate `conn.params` with:
 
       %{"rest" => ["value", "extra"]}
-
-  # Options
-
-  The router accepts a list of options:
-
-    * `:known_methods` - custom list of HTTP methods known by your
-      server, for example: `["GET", "HEAD", "OPTIONS", "TRACE"]`
-
-  This option will override the default list of methods that each
-  Resource knows about (GET, HEAD, POST, PUT, PATCH, DELETE, and
-  OPTIONS). If a Resource implements the `known_methods` callback,
-  that list always takes precedence over the default list.
   """
 
 
   @doc false
-  defmacro __using__(options) do
+  defmacro __using__(_options) do
     quote location: :keep do
       use Plug.Router
       import PlugRest.Router
       @before_compile PlugRest.Router
-
-      @known_methods Keyword.get(unquote(options), :known_methods)
     end
   end
 
@@ -178,8 +164,6 @@ defmodule PlugRest.Router do
             false ->
               unquote(options)
           end
-
-        options = Keyword.put_new(options, :known_methods, @known_methods)
 
         case function_exported?(unquote(handler), :call, 1) do
           true ->